feat: add services

+++ src/db/supabase/services/problem.service.ts
@@ -2,12 +2,12 @@
 
 import { createClient } from '@/db/supabase/server'
 import { Problem } from '@/types/table.type'
-import { CreateProblemInput } from '@/types/dto/problem.dto'
+import { CreateProblemDto, CreateProblemInput } from '@/types/dto/problem.dto'
 
-export const getProblems = async (): Promise<Problem[]> => {
-  const supabase = await createClient()
-
+export const getProblems = async (): Promise<Problem[] | undefined> => {
   try {
+    const supabase = await createClient()
+
     const { data, error } = await supabase
       .from('problems')
       .select('*')
@@ -15,20 +15,21 @@ export const getProblems = async (): Promise<Problem[]> => {
 
     if (error) {
       console.error('getProblems error:', error)
-      throw error
+      return []
     }
 
     return data
   } catch (error) {
     console.error('getProblems error:', error)
-    throw error
   }
 }
 
-export const getProblem = async (pNumber: number): Promise<Problem> => {
-  const supabase = await createClient()
-
+export const getProblem = async (
+  pNumber: number
+): Promise<Problem | undefined> => {
   try {
+    const supabase = await createClient()
+
     const { data, error } = await supabase
       .from('problems')
       .select('*')
@@ -37,43 +38,51 @@ export const getProblem = async (pNumber: number): Promise<Problem> => {
 
     if (error) {
       console.error('getProblem error:', error)
-      throw error
+      return undefined
     }
 
     return data
   } catch (error) {
     console.error('getProblem error:', error)
-    throw error
   }
 }
 
 export const createProblem = async (
   problem: CreateProblemInput
-): Promise<Problem> => {
-  const supabase = await createClient()
-
+): Promise<Problem | undefined> => {
   try {
+    const validatedProblem = CreateProblemDto.parse(problem)
+
+    const supabase = await createClient()
+
     const { data, error } = await supabase
       .from('problems')
-      .insert([problem])
+      .insert([validatedProblem])
+      .select()
       .single()
 
     if (error) {
       console.error('createProblem error:', error)
-      throw error
+      return undefined
     }
 
     return data
   } catch (error) {
     console.error('createProblem error:', error)
-    throw error
   }
 }
 
-export const deleteProblem = async (pNumber: number): Promise<Problem> => {
-  const supabase = await createClient()
+export const deleteProblem = async (
+  pNumber: number
+): Promise<Problem | undefined> => {
+  if (!pNumber) {
+    console.error('pNumber is required')
+    return undefined
+  }
 
   try {
+    const supabase = await createClient()
+
     const { data, error } = await supabase
       .from('problems')
       .delete()
@@ -82,12 +91,11 @@ export const deleteProblem = async (pNumber: number): Promise<Problem> => {
 
     if (error) {
       console.error('deleteProblem error:', error)
-      throw error
+      return undefined
     }
 
     return data
   } catch (error) {
     console.error('deleteProblem error:', error)
-    throw error
   }
 }
